Do not sync keyboard state on keyDown events.
This causes all kinds of issues because sying the state like this resets any held keys.

This means capslock/numlock on the RDP client can be out of sync from your system, but
this has been how it has worked in the past.

This can be made to work, I think, by re-sending keyDown events for any held key every
time we send a synchronize event - but that is a somewhat non-trivial change.